Feature/sync review (#29)
* fix review methods

add reference to Review & Item entity
separate review controller from item's
add method synchronizing review & member & item entity
add method calculating ratings & num of reviews

* Synchronizing Item Review Count and Ratings

add reference valid items and members to review
add calculation for Review counts and ratings

+++ src/main/java/test/shop/application/dto/request/ProductDto.java
@@ -19,7 +19,7 @@ public class ProductDto{
     private int price;
     private int stockQuantity;
     private String description;
-    private int ratings;
+    private double ratings;
     private int numOfReviews;
     private Category category;
     private String thumbnailUrl;
